requirements.txt
discord.py==2.5.2
aiohttp==3.12.13
python-dotenv==1.1.1
dateparser==1.2.2
